OSINT APP/requirements.txt
##########################Requirements########################################################
tldextract
dnspython

